\name{balance}
\alias{balance}
\docType{data}
\title{Balance Data}

\description{
 Scores of 484 todlers on 25 balance-problem items.
}

\usage{
data(balance)
}

\format{
A 484 by 25 data frame containing integers. 
}

\details{
 The items include 5 conflict-balance (CB) items , 5 conflic-distance
 (CD) items, 5 conflict-weight (CW) items, 5 distance (D) items, and 5 weight (W) items.
 Respondents have been deidentified, and covariates have been removed. 
 The deidentified data do allow to replicate the analyses in Sijtsma and Van der Ark (2020, chapter 5) using the code available from \url{https://osf.io/e9jrz}. 
}

\source{
 The data were collected by Leo van Maanen (see, Van Maanen, Been & Sijtsma, 1989). 
}

\references{

Van Maanen, L., Been, P. H., & Sijtsma, K. (1989). 
Problem solving strategies and the Linear Logistic Test Model. 
In E. E. Ch. I. Roskam (Ed.), \emph{Mathematical psychology in progress} (pp. 267-287). 
Springer.

Sijtsma, K., & Van der Ark, L. A. (2020),
\emph{Measurement models for psychological attributes.} Chapman and Hall/CRC Press.
\url{https://www.routledge.com/Measurement-Models-for-Psychological-Attributes/Sijtsma-Ark/p/book/9780367424527}

}

\examples{
data(balance)
}

\keyword{datasets}
