%to do: more examples

\name{DS14B}
\alias{DS14B}
\docType{data}
\title{DS14 Bootstrap sample}
\description{
DS14. Not the real data set used in Straat et al. (2016) but a bootstrap sample.
}

\usage{data(DS14B)}
\format{
  A data frame with 3111 bootstrapped observations on 14 variables.
}

\references{
 Straat, J. H., van der Ark, L. A., and Sijtsma, K. (2016).
 Using conditional association to identify local dependence.
 Manuscript submitted for publication.
}

\examples{
data(DS14B)
}
\keyword{datasets}
